# Fix GH-11240: PHP crashes on big file with array inside