@dstogov 2023-05-22
Also see an implementation of improved algorithm https://github.com/DavePearce/StronglyConnectedComponents