@nielsdos 2023-05-26
I checked out https://arxiv.org/pdf/1703.10023.pdf. It does provide some optimisation ideas, one of which already happens (combine visited[] and dfs[] into a single array dfs). And it also gives some more ideas. It does not however provide a reference algorithm for iterative SCC.

> Maintaining of stack with 3 additional lists looks a bit weird to me.
> Did you use some reference algorithm implementation or invent this yourself?

I did it myself. Let me explain the additional items on the stack.
It's not strictly necessary to push the items on the stack, this is just a performance improvement.
The reason: when we return to a var on the recursion stack, we want to resume the iteration of the uses from the last use or phi node that caused the recursion. Iterating from the start of the use list is also correct, but wasteful of performance: it just updates the root but since the root was already updated this doesn't do anything new.

In Dave Pearce's algorithm we'd still have to maintain the additional items on the stack.
Dave Pearce does however improve the space cost of Tarjan's algorithm by merging two lists into one. I can probably take a look at implementing this if you want.