@dstogov 2023-10-27
Closed in favour of https://github.com/php/php-src/pull/12528